Fixed unit tests and retabed ROVppAS to spaces (tab width 4).
Had to add a check for policy vector to be greater than size 0 in
the send_all_announcmenet function, because some tests don't initizlize
the
policy_vector. For those I could identify from the error I added to the
test
but those I couldn't identify I just added a check for policy_vector
size
being greater than 0.

Had to update the pass_rov test because it was not designed
to work with roa_validity.

best_alternative_ann_chosen test was broken before and now is using
accurately checking that the actual best announcement is being selected.

Other unit test fixes had to add consideration for roa_validity and/or
adding initializing policy_vector.