Fix issue 15
Location: dead code
Bug: In an if else chain, there were debug statements that were
commented out. However, when run, this dead code actually got called in
the process announcements function, meaning the dead code got run
instead of the standard process announcements
Expected Impacted: V3 completely fails because of this, possibly other
policies as well. I believe this would also have affected our results.
However, this is only true if calling this "dead code" was actually
intentional. This gets called in effort to "move the entire prefix" in
the same direction (i.e. received_from_asn) as an issued
preventive_announcement. If recalled well, this idea was retracted. If
that's the case, then it's working correctly (albiet in a terribly
confusing way).
Discussed: No

Removed the override parameter from the defintion of
process_announcement and removed any uses of it from ROVppAS.cpp.
It seems that the only place it was used was in V3 to effectively do
nothing (i.e. not do some idea we had before without removing the
functionality to add it back easily).