Fix for Issue_ID 13 really fixed
The problem with the prior commit for the fix for Issue_ID 13 was
it didn't consider the consequence of adding security considerat-
ions to process_announcement. This messed up BGP and ROV, because
now they were also doing some ROVpp checks.

That caused some tests to fail (e.g. propagate_up and propagate_down).
Now BGP and ROV don't do these checks.

In the process I also fixed some other tests which needed policy_vector
to be added to the test.

There was a test called best_alternative_route2, which was developed
to ensure ROVpp picks relationship over security. This test was not made
to properly test the fix for issue 13 because it checks the outcome of
best_alternative_route directly rather than checking the resulting
loc_rib, which also needs to go through the process_announcement
function to have a correct outcome to check. Now the test has been
updated and the test passes!