Fixed a few issues (which turns out to be just one we already solved) and renamed a test
The following is verbatim from issue description in Slack (Cameron
Identified):

```
not sure if it came from the recent fixes, but I don't think
ROVppAS::within_bad_neighbor_hole works the way we want it to for
checking the safety of the path. It should check on line 119 if the bad
ann.prefix is contained in or equal to ann.prefix, not the other way
around. In the subprefix hijack case, the hole will always be smaller
than the legitimate announcement.

also relating to that check in ROVppAS::process_announcements lines
442-445 and similar sections, are we sure it's correct to only process
the announcement if it is not with the bad neighbor hole? If it passes
the tests, then I'm okay with it, but it looks concerningly similar to
the original bad_neighbors bug where where the ROV++ ASes were rejecting
all announcements from bad neighbors.

in ROVppTest.cpp, test_best_alternative_route2 no longer checks the
best_alternative_route, we should rename it to reflect the function it
tests. Before we do that, though, that unit test was added specifically
to test for a fix of issue 13, and the way that comment is written
indicates that the best_alternative_route function will still
incorrectly choose a3 instead of a1, which might be a problem.
```

We have addressed the above in a meeting. Cameron made the fixes.
Funny part is the bugs we solved collapsed to one bug we accidently
re-introduced, but now is hopefully resolved for good!