Fix for Issue_ID 17
Location: Process_announcement
Bug: Extrapolator fails for system tests for bug_scenario_sys_test,
figures 3, 4 and withdraws
Cause: From fix for Issue_ID 12, a bug was introduced that did not
properly separate BGP and others to not execute ROVpp considerations.
Expected Impacted: None as this bug was introduced from a fix

To fix had to fix some logic statement for proper separation from
ROVpp policies and others.