Fix issue 17 (#70)
* Fixed issue 17, but needs testing and cleaning

Issue_ID: 17
Location: Unknown, probably best_alternative_announcement
Location (Updated): process_announcement (singular)
Bug: Propagation graph does not converge in a reasonable amount of time (possible route flapping?) in non-lite versions (only tested v1 and v1 lite). It’s possible that this is correct behavior, but we don't have a route-flapping test scenario (we don't have one at all actually) that validates this.
Bug (Updated):  Condition to select best announcement in process_announcement used an inaccurate assumption about new announcement being processed, which caused route flapping between announcements with and without holes.
Cause: unknown (something specific to non-lite functions)
Cause (Updated): incorrect assumption about new announcement being processed in process_announcement. The assumption was that the new announcement is safe, but that's not exactly true, because the method to determine safety was not accurate enough when it was submitted for processing (i.e. in process_announcements we process announcements that pass_rov without additional checking, which leads to considering announcements that have holes).
Expected Impact: unknown
Expected Impact (Updated): route flapping

* Removed the debugging print statements and tested all system tests to make sure they complete within 1 or 2 propagation cycles.

* Actually these latest commits have been for Issue_ID 18.
This particular commit adds a consideration for the case
in which both the current and the new announcement are not
safe. In which case the priority is then checked to pick best.

* Added fix for exceeding 100 propagation cycles
Had to handle additional cases which were not
considered.

* Removed debugging print statements.

Co-authored-by: jfuruness <jfuruness@gmail.com>