Fixed issue 14 (#73)
* Fixed issue 14

Issue Id: 14
Location: Unknown
Bug: Fixes of other bugs reintroduced routing loops in v3
Cause: turns out the ATTACKER_ON_ROUTE_FLAG was actually a needed flag for preventive announcements to prevent the kind of loop we're noticing now. This is noted in the loop documentation we have on overleaf.
Expected Impacted: need to fix before getting additional v3 and v3 lite results
Discussed: Yes

Had to add the ATTACKER_ON_ROUTE_FLAG back in, but this time properly impelemented.
It solves the loop scenario which is a saved full sim run.

* Update ROVppAS.cpp

Nice! Good to delete!

Co-authored-by: Cameron Morris <cameron.morris@uconn.edu>

Co-authored-by: jfuruness <jfuruness@gmail.com>
Co-authored-by: Cameron Morris <cameron.morris@uconn.edu>