Fixed issue 20 (#76)
* Fix for Issue_ID 20

Removed the 'drop attacks from customers' feature for v2, because it's causing hidden hijacks.

* Fix Issue_ID 20

Issue Id: 20
Location: ROVppAS::process_anouncements, ROVppExtrapolator::send_all_announcements
Bug: Fixes of other bugs impaired the performance of v2 lite and potentially v2 (noticable in trace_hijacked_adopting)
Cause: This is caused by the special case when an attack annoucement is received from a customer. In this case we should drop the announcement and blackhole. Before it was just dropping the announcement, which created a hidden hijacks.
Expected Impacted: hijack rate is higher than all other ROV++ policies during first 60%
Discussed: Yes

* Applying better implementation from review

In review we noticed a better way of checking if announcement was
from a customer. That is check the priority. In that case the
share_special_ann flag in Announcement.h was removed and the
corresponding code supporting it.

Co-authored-by: jfuruness <jfuruness@gmail.com>
Co-authored-by: Reynaldo Morillo <reymorillo2@gmail.com>