Fixes for Issue_ID 20
Turns out that making copies of the announcements for the blackholes
variable was not a good idea, because later on in the code, there's
a potentially in-adventent dependency that the announcement is a
reference rather than a copy (that is in is_filtered function in
ROVppExtrapolator checked the origin of the announcement to determine
if it's part of the blackhole list). Also corrected the logic for sending
of blackhole announcements to cutomers, provider and peers.

The code in this state passes all the system tests.